requirements.txt
alembic==1.12.1
bcrypt==4.0.1
certifi==2023.7.22
greenlet==3.0.1
Mako==1.3.0
MarkupSafe==2.1.3
psycopg2==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
sentry-sdk==1.35.0
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
urllib3==2.1.0
